feat: move mcp show/kill to db ps/kill
These commands are about database connections, not MCP. Renamed
show -> ps and moved to the db command group.
